Change class name of the mixin
