+++ plexapi/audio.py
@@ -4,6 +4,7 @@ from urllib.parse import quote_plus
 from plexapi import library, media, utils
 from plexapi.base import Playable, PlexPartialObject
 from plexapi.exceptions import BadRequest
+from plexapi.mixins import SplitMergeAble
 
 
 class Audio(PlexPartialObject):
@@ -123,7 +124,7 @@ class Audio(PlexPartialObject):
 
 
 @utils.registerPlexObject
-class Artist(Audio):
+class Artist(Audio, SplitMergeAble):
     """ Represents a single Artist.
 
         Attributes:
