+++ plexapi/video.py
@@ -5,6 +5,7 @@ from urllib.parse import quote_plus, urlencode
 from plexapi import library, media, settings, utils
 from plexapi.base import Playable, PlexPartialObject
 from plexapi.exceptions import BadRequest, NotFound
+from plexapi.mixins import SplitMergeAble
 
 
 class Video(PlexPartialObject):
@@ -259,7 +260,7 @@ class Video(PlexPartialObject):
 
 
 @utils.registerPlexObject
-class Movie(Playable, Video):
+class Movie(Playable, Video, SplitMergeAble):
     """ Represents a single Movie.
 
         Attributes:
@@ -383,7 +384,7 @@ class Movie(Playable, Video):
 
 
 @utils.registerPlexObject
-class Show(Video):
+class Show(Video, SplitMergeAble):
     """ Represents a single Show (including all seasons and episodes).
 
         Attributes:
